Remove duplicate hook.
